fix(db): make map tags migration have proper field type

+++ libs/db/src/migrations/20241220134651_map_tags/migration.sql
@@ -6,4 +6,4 @@
 */
 -- AlterTable
 ALTER TABLE "Leaderboard" DROP COLUMN "tags",
-ADD COLUMN     "tags" INTEGER[];
+ADD COLUMN     "tags" SMALLINT[];
